Add code to print build_stdout after build is finished (#103)
* Add code to print build_stdout after build is finished
Build_stdout of all ABIs is always updated whenever built. And the ABI
in which the device is not connected or failed to perform the unit test
does not include "Gtest Success". Therefore, build_stdout is printed only
when "Gtest Success" is included.

* Modify to reflect code review
Now, if the device does not exist, it will print the result that the
unit test was not run.

Co-authored-by: yeonns <yeonns@users.noreply.github.com>